src/requirements.txt
auto_mix_prep==0.2.0
html2text==2020.1.16
matplotlib==3.5.2
numpy==1.22.4
PyQt5==5.15.6
PyYAML==6.0
requests==2.27.1
